+++ low_pass_filter/src/window_functions.h
@@ -3,8 +3,8 @@
 
 #include <math.h>
 
-void bartlett_window(float* coeffs, int num_coeffs);
-void blackman_window(float* coeffs, int num_coeffs);
-void hamming_window(float* coeffs, int num_coeffs);
-void hanning_window(float* coeffs, int num_coeffs);
-void kaiser_window(float* coeffs, int num_coeffs);
+void bartlett_window(float* coeffs, size_t num_coeffs);
+void blackman_window(float* coeffs, size_t num_coeffs);
+void hamming_window(float* coeffs, size_t num_coeffs);
+void hanning_window(float* coeffs, size_t num_coeffs);
+void kaiser_window(float* coeffs, size_t num_coeffs);
